Move Singleton class to own package
