+++ src/main/java/creational/singleton/Singleton.java
@@ -1,4 +1,4 @@
-package creational;
+package creational.singleton;
 
 public class Singleton {
     private static volatile Singleton instance;
